requirements.txt
pyyaml
dash
dash-renderer
dash-html-components
dash-core-components
plotly
fire
ujson